+++ routers/router.go
@@ -8,12 +8,15 @@
 package routers
 
 import (
-	"github.com/udistrital/sga_plan_estudio_mid/controllers"
-
 	"github.com/astaxie/beego"
+	"github.com/udistrital/sga_plan_estudio_mid/controllers"
+	"github.com/udistrital/utils_oas/errorhandler"
 )
 
 func init() {
+
+	beego.ErrorController(&errorhandler.ErrorHandlerController{})
+
 	ns := beego.NewNamespace("/v1",
 		beego.NSNamespace("/plan-estudios",
 			beego.NSInclude(
